Restraining access (PUT, DELETE)

+++ src/http/server/modules/methods/DeleteRequest.java
@@ -7,16 +7,28 @@ import http.server.modules.header.ResponseHttpHeader;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.OutputStream;
+import java.nio.file.Path;
 
 public class DeleteRequest implements Method {
     @Override
     public void processMethod(HttpHeader header, BufferedReader input, OutputStream outputStream) {
-        String path = "resources" + header.getResource();
-        File f = new File(path);
+        Path path = Path.of("resources" + header.getResource()).toAbsolutePath().normalize();
+        Path pRestraint = Path.of("resources/uploads").toAbsolutePath().normalize();
 
-        f.delete();
+        HttpStatusCode httpStatusCode;
+        if(path.startsWith(pRestraint))
+        {
+            File f = path.toFile();
 
-        ResponseHttpHeader responseHttpHeader = new ResponseHttpHeader(HttpStatusCode.ACCEPTED_202);
+            f.delete();
+            httpStatusCode = HttpStatusCode.ACCEPTED_202;
+        }
+        else
+        {
+            httpStatusCode = HttpStatusCode.FORBIDDEN_403;
+        }
+
+        ResponseHttpHeader responseHttpHeader = new ResponseHttpHeader(httpStatusCode);
         responseHttpHeader.write(outputStream);
     }
 }
